Add files via upload
The project was born out of a small and public availability 
of a dataset from Kaggle, which included a useful dataset of 
real world self-reported feelings of symptoms (less than 2000 
records). Relying on such a small sample only for training of
a model would bring a huge risk of overfitting.To overcome this we used a synthetic data generation 
technique. Statistical distributions and correlations of the 
original sample were examined and used to create a larger 
synthetic dataset of 9001 unique records.